// js-packages/openk9-admin-new/src/graphql-generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInteger: any;
  DateTime: any;
};

export type AclMapping = {
  __typename?: 'AclMapping';
  docTypeField?: Maybe<DocTypeField>;
  key?: Maybe<PluginDriverDocTypeFieldKey>;
  pluginDriver?: Maybe<PluginDriver>;
  userField?: Maybe<UserField>;
};

export type Analyzer = {
  __typename?: 'Analyzer';
  charFilters?: Maybe<Connection_CharFilter>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  jsonConfig?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  tokenFilters?: Maybe<Connection_TokenFilter>;
  tokenizer?: Maybe<Tokenizer>;
  type?: Maybe<Scalars['String']>;
};
// ...
export type Annotator = {
  __typename?: 'Annotator';
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  docTypeField?: Maybe<DocTypeField>;
  docTypeFieldNotInAnnotator?: Maybe<Connection_DocTypeField>;
  extraParams?: Maybe<Scalars['String']>;
  fieldName?: Maybe<Scalars['String']>;
  fuziness?: Maybe<Fuzziness>;
  id?: Maybe<Scalars['ID']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  size?: Maybe<Scalars['Int']>;
  type?: Maybe<AnnotatorType>;
};
// ...
export type Bucket = {
  __typename?: 'Bucket';
  catIndices?: Maybe<Array<Maybe<CatResponse>>>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  datasources?: Maybe<Connection_Datasource>;
  description?: Maybe<Scalars['String']>;
  docCount?: Maybe<Scalars['BigInteger']>;
  enabled: Scalars['Boolean'];
  id?: Maybe<Scalars['ID']>;
  indexCount?: Maybe<Scalars['BigInteger']>;
  language?: Maybe<Language>;
  languages?: Maybe<Connection_Language>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  queryAnalysis?: Maybe<QueryAnalysis>;
  ragConfigurationChat?: Maybe<RagConfiguration>;
  ragConfigurationChatTool?: Maybe<RagConfiguration>;
  ragConfigurationSimpleGenerate?: Maybe<RagConfiguration>;
  refreshOnDate?: Maybe<Scalars['Boolean']>;
  refreshOnQuery?: Maybe<Scalars['Boolean']>;
  refreshOnSuggestionCategory?: Maybe<Scalars['Boolean']>;
  refreshOnTab?: Maybe<Scalars['Boolean']>;
  retrieveType?: Maybe<RetrieveType>;
  searchConfig?: Maybe<SearchConfig>;
  sortings?: Maybe<Connection_Sorting>;
  suggestionCategories?: Maybe<Connection_SuggestionCategory>;
  tabs?: Maybe<Connection_Tab>;
};
// ...
export type Connection_Analyzer = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_Analyzer>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_Annotator = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_Annotator>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_Bucket = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_Bucket>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};

/** A connection to a list of items. */
export type Connection_CharFilter = {
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<Edge_CharFilter>>>;
  /** details about this specific page */
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type DataIndex = {
  __typename?: 'DataIndex';
  cat?: Maybe<CatResponse>;
  chunkType?: Maybe<ChunkType>;
  chunkWindowSize?: Maybe<Scalars['Int']>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  datasource?: Maybe<Datasource>;
  description?: Maybe<Scalars['String']>;
  docCount?: Maybe<Scalars['BigInteger']>;
  docTypes?: Maybe<Connection_DocType>;
  embeddingDocTypeField?: Maybe<DocTypeField>;
  embeddingJsonConfig?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  knnIndex?: Maybe<Scalars['Boolean']>;
  mappings?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  settings?: Maybe<Scalars['String']>;
};
// ...
export type Datasource = {
  __typename?: 'Datasource';
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  dataIndex?: Maybe<DataIndex>;
  dataIndexes?: Maybe<Connection_DataIndex>;
  description?: Maybe<Scalars['String']>;
  enrichPipeline?: Maybe<EnrichPipeline>;
  id?: Maybe<Scalars['ID']>;
  jsonConfig?: Maybe<Scalars['String']>;
  /** Last ingestion date of data for current datasource (ISO-8601) */
  lastIngestionDate?: Maybe<Scalars['DateTime']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  pluginDriver?: Maybe<PluginDriver>;
  /** The duration to identify orphaned Dataindex. */
  purgeMaxAge?: Maybe<Scalars['String']>;
  /** If true set active the purge job scheduling */
  purgeable?: Maybe<Scalars['Boolean']>;
  /** Chron quartz expression to define purging for this datasource */
  purging?: Maybe<Scalars['String']>;
  /** If true set datasource as reindexable */
  reindexable?: Maybe<Scalars['Boolean']>;
  /** Chron quartz expression to define reindexing of datasource */
  reindexing?: Maybe<Scalars['String']>;
  /** If true set datasource as schedulable */
  schedulable?: Maybe<Scalars['Boolean']>;
  schedulers?: Maybe<Connection_Scheduler>;
  /** Chron quartz expression to define scheduling of datasource */
  scheduling?: Maybe<Scalars['String']>;
};
// ...
export type DefaultConnection_DataIndex = Connection_DataIndex & {
  __typename?: 'DefaultConnection_DataIndex';
  edges?: Maybe<Array<Maybe<Edge_DataIndex>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type DefaultEdge_DataIndex = Edge_DataIndex & {
  __typename?: 'DefaultEdge_DataIndex';
  cursor?: Maybe<Scalars['String']>;
  node?: Maybe<DataIndex>;
};
// ...
export type EnrichPipeline = {
  __typename?: 'EnrichPipeline';
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  enrichItems?: Maybe<Connection_EnrichItem>;
  id?: Maybe<Scalars['ID']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
};
// ...
export type PluginDriver = {
  __typename?: 'PluginDriver';
  aclMappings?: Maybe<Array<Maybe<PluginDriverAclMapping>>>;
  /** ISO-8601 */
  createDate?: Maybe<Scalars['DateTime']>;
  description?: Maybe<Scalars['String']>;
  docTypeFields?: Maybe<Connection_DocTypeField>;
  id?: Maybe<Scalars['ID']>;
  jsonConfig?: Maybe<Scalars['String']>;
  /** ISO-8601 */
  modifiedDate?: Maybe<Scalars['DateTime']>;
  name?: Maybe<Scalars['String']>;
  provisioning?: Maybe<Provisioning>;
  type?: Maybe<PluginDriverType>;
};
// ...
export enum Provisioning {
  System = 'SYSTEM',
  User = 'USER'
}

/** Query root */
export type Query = {
  __typename?: 'Query';
  analyzer?: Maybe<Analyzer>;
  analyzers?: Maybe<Connection_Analyzer>;
  annotator?: Maybe<Annotator>;
  annotators?: Maybe<Connection_Annotator>;
  bucket?: Maybe<Bucket>;
  buckets?: Maybe<Connection_Bucket>;
  charFilter?: Maybe<CharFilter>;
  charFilters?: Maybe<Connection_CharFilter>;
  dataIndex?: Maybe<DataIndex>;
  dataIndices?: Maybe<Connection_DataIndex>;
  datasource?: Maybe<Datasource>;
  datasources?: Maybe<Connection_Datasource>;
  docType?: Maybe<DocType>;
  docTypeField?: Maybe<DocTypeField>;
  docTypeFieldNotInAnnotator?: Maybe<Connection_DocTypeField>;
  docTypeFields?: Maybe<Connection_DocTypeField>;
  docTypeFieldsByParent?: Maybe<Connection_DocTypeField>;
  docTypeFieldsFromDocType?: Maybe<Connection_DocTypeField>;
  docTypeFieldsFromDocTypeByParent?: Maybe<Connection_DocTypeField>;
  docTypeFieldsNotInSorting?: Maybe<Connection_DocTypeField>;
  docTypeFieldsNotInTokenTab?: Maybe<Connection_DocTypeField>;
  docTypeTemplate?: Maybe<DocTypeTemplate>;
  docTypeTemplates?: Maybe<Connection_DocTypeTemplate>;
  docTypes?: Maybe<Connection_DocType>;
  embeddingModel?: Maybe<EmbeddingModel>;
  embeddingModels?: Maybe<Connection_EmbeddingModel>;
  enabledBucket?: Maybe<Bucket>;
  enrichItem?: Maybe<EnrichItem>;
  enrichItems?: Maybe<Connection_EnrichItem>;
  enrichPipeline?: Maybe<EnrichPipeline>;
  enrichPipelines?: Maybe<Connection_EnrichPipeline>;
  /** Returns the list of events */
  event?: Maybe<Array<Maybe<Event>>>;
  /** Get event data */
  eventData?: Maybe<Scalars['String']>;
  /** Returns the list of available options for the event */
  eventOptions?: Maybe<Array<Maybe<EventOption>>>;
  language?: Maybe<Language>;
  languages?: Maybe<Connection_Language>;
  largeLanguageModel?: Maybe<LargeLanguageModel>;
  largeLanguageModels?: Maybe<Connection_LargeLanguageModel>;
  pluginDriver?: Maybe<PluginDriver>;
  pluginDrivers?: Maybe<Connection_PluginDriver>;
  pluginDriversPageFilter?: Maybe<Page_PluginDriver>;
  queryAnalyses?: Maybe<Connection_QueryAnalysis>;
  queryAnalysis?: Maybe<QueryAnalysis>;
  queryParserConfig?: Maybe<QueryParserConfig>;
  queryParserConfigs?: Maybe<Connection_QueryParserConfig>;
  ragConfiguration?: Maybe<RagConfiguration>;
  ragConfigurations?: Maybe<Connection_RagConfiguration>;
  rule?: Maybe<Rule>;
  rules?: Maybe<Connection_Rule>;
  scheduler?: Maybe<Scheduler>;
  schedulers?: Maybe<Connection_Scheduler>;
  searchConfig?: Maybe<SearchConfig>;
  searchConfigs?: Maybe<Connection_SearchConfig>;
  sorting?: Maybe<Sorting>;
  sortings?: Maybe<Connection_Sorting>;
  suggestionCategories?: Maybe<Connection_SuggestionCategory>;
  suggestionCategory?: Maybe<SuggestionCategory>;
  tab?: Maybe<Tab>;
  tabs?: Maybe<Connection_Tab>;
  tokenFilter?: Maybe<TokenFilter>;
  tokenFilters?: Maybe<Connection_TokenFilter>;
  tokenTab?: Maybe<TokenTab>;
  tokenTabs?: Maybe<Connection_TokenTab>;
  tokenizer?: Maybe<Tokenizer>;
  tokenizers?: Maybe<Connection_Tokenizer>;
  totalSortings?: Maybe<Connection_Sorting>;
  totalTokenTabs?: Maybe<Connection_TokenTab>;
  unboundAnalyzersByCharFilter?: Maybe<Array<Maybe<Analyzer>>>;
  unboundAnalyzersByTokenFilter?: Maybe<Array<Maybe<Analyzer>>>;
  unboundBucketsByDatasource?: Maybe<Array<Maybe<Bucket>>>;
  unboundBucketsBySuggestionCategory?: Maybe<Array<Maybe<Bucket>>>;
  unboundBucketsByTab?: Maybe<Array<Maybe<Bucket>>>;
  /** Fetches DocTypeFields unbound to the provided SuggestionCategory ID with FieldType KEYWORD or I18N. */
  unboundDocTypeFieldsBySuggestionCategory?: Maybe<Array<Maybe<DocTypeField>>>;
  unboundEnrichPipelines?: Maybe<Array<Maybe<EnrichPipeline>>>;
  /**
   * Retrieves a list of RAGConfiguration entities of the specified RAGType
   * that are not yet associated with the given Bucket.
   *
   * This query returns all RAGConfiguration instances that:
   * - Have the specified RAGType.
   * - Are not currently linked to the provided bucketId.
   *
   * Arguments:
   * - `bucketId` (ID!): The ID of the Bucket for which to retrieve unbound RAGConfiguration entities.
   * - `ragType` (RAGType!): The type of RAGConfiguration to filter by.
   *
   * Returns:
   * - A list of RAGConfiguration entities that match the criteria.
   *
   */
  unboundRAGConfigurationByBucket?: Maybe<Array<Maybe<RagConfiguration>>>;
  unboundTabsByTokenTab?: Maybe<Array<Maybe<Tab>>>;
};
// ...
export type DataSourceQuery = { __typename?: 'Query', datasource?: { __typename?: 'Datasource', id?: string | null, name?: string | null, description?: string | null, schedulable?: boolean | null, scheduling?: string | null, jsonConfig?: string | null, reindexable?: boolean | null, reindexing?: string | null, purgeable?: boolean | null, purging?: string | null, purgeMaxAge?: string | null, lastIngestionDate?: any | null, pluginDriver?: { __typename?: 'PluginDriver', id?: string | null, name?: string | null } | null, dataIndex?: { __typename?: 'DataIndex', id?: string | null, name?: string | null, description?: string | null, knnIndex?: boolean | null } | null, enrichPipeline?: { __typename?: 'EnrichPipeline', id?: string | null, name?: string | null } | null, dataIndexes?: { __typename?: 'DefaultConnection_DataIndex', edges?: Array<{ __typename?: 'DefaultEdge_DataIndex', node?: { __typename?: 'DataIndex', id?: string | null, name?: string | null } | null } | null> | null } | null } | null };